Initial commit
I started out trying to make a dropped ice cream cone, but it became
problematic due to the orthogonal nature of the ellipse. I was
disastisfied with the messy tangents produced by connecting the cone's
triangle to the ice cream's ellipse. If we could rotate, things might
line up better, but even with the grounding plane produced by the
rectangle, I'm not sure the angles would be convincing.